+++ packages/react-noop-renderer/src/ReactNoopServer.js
@@ -14,11 +14,11 @@
  * environment.
  */
 
-import ReactFizzStreamer from 'react-server';
+import ReactFizzServer from 'react-server';
 
 type Destination = Array<string>;
 
-const ReactNoopServer = ReactFizzStreamer({
+const ReactNoopServer = ReactFizzServer({
   scheduleWork(callback: () => void) {
     callback();
   },
