+++ packages/react-server/index.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from './src/ReactFizzStreamer';
+export * from './src/ReactFizzServer';
